+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - library item "more" actions: re-scan, remove progress, mark complete, and remove from continue listening
 - more actions can also be triggered by long press of item card on home screen
 - playback history tile now shows chapter name alongside position
+- time adjusted per speed setting, chapter/total progress bar toggle
 
 ## [v0.5.14] - 2026-08-01